Simplified the code for SWSender in stopAndWait.js
Changed stopAndWaitTest.js to expect SWSender not to
use the property currentState, but to instead use new
properties currentSeqNum and isWaitingAck for maintaining
the state of an instance.

From SWSender, removed the constants WAIT_CALL_0, WAIT_CALL_1,
WAIT_ACK_0 and WAIT_ACK_0, the property currentState,
and the private methods _waitCall() and _waitAck().
On the other hand, new properties were added to SWSender,
including currentSeqNum and isWaitingAck, maintaining the
same previous functionality.

Updated the documentation accordingly.