Added a new property called ackNum to Packet
Changed the test description for some default properties
of Packet and added a test for the default value of ackNum
int PacketTest.js
Changed the tests in stopAndWait.js to now use ackNum to get
the acknowledged sequence number.
Changed SWSender and SWReceiver to pass the changed tests.
Updated the documentation with the new property.