Added a way for channels to lose packets based on probability
Added a test for the new functionality, wich involves giving
a loss probability to the Channel constructor. Changed the
implementation of send() in Channel to compute whether
a packet is going to be lost or not and pass the new test.
Also, added a method onSend() to Channel, fired when a packet
is sent by send(). On the other hand, changed losePacket() to return
whether the packet was traveling or not. And finally, changed
the documentation accordingly.