Added a way for channels to damage packets based on probability
Added a test for the new functionality, that involves passing a
damage probability to the constructor.
Changed the send() method to pass the new test.
Updated the documentation for the new option of the constructor.
Changed onSend(), adding a new parameter to know if the packet
is going to be lost.