Added and refactored tests in stopAndWaitTest.js
Refactored all the tests in stopAndWait.js to use beforeEach
where it could be done. And added new tests for Stop-and-Wait,
including one where SWSender and SWReceiver act together.
Did some minor changes to SWSender, so it could pass all the
tests and changed its documentation for one of the default
parameters (timeoutTime).