Added new files, tests and classes for Go-Back-N
Added a goBackNTest.js file that has definitions for testing
the new classes for Go-Back-N. For now the file only has
implementations of tests for the sender of GBN.
Added class definitions for GBN, with one of them implementing
the sender class, GBNSender, to pass the previous tests.
Added documentation for the new classes and the implemented
methods.