Changed the send method of Channel
Changed the tests in ChannelTest.js to expect send() to return a
boolean, and therefore removed the expectation for a promise.
Changed Channel.js to pass the tests, adding new callback methods
onPacketLost and onPacketStopped.
Also changed property travelingPackets to "private" and replaced
it with a method getTravelingPackets(), that returns an interator.
Changed the documentation accordingly.

Finally, the send() test for Node is not passing due to the changes.