Added a test for stop-and-wait and changed SWSender
Added a test in stopAndWaitTest.js that expects the send method
of SWSender to return true if the sender is not waiting for an
acknowledgment and can send a packet, and false if it can not send
a packet.
Changed SWSender to pass the new test.
Changed the documentation accordingly.