Added a new test for go-back-n and changed GBNSender
Added a new test in goBackNTest.js that expects the send() method
of GBNSender to return a boolean that is true if a packet could
be sent because the window had space and is false if not.
Changed send() from GBNSender to pass the new test.
Changed the documentation of GBNSender.