Added tests for selective-repeat
Added definitions of tests for selective-repeat in a new file
called selectiveRepeatTest.js.
Implemented the tests for send() and stop(), for the sender
class of selective-repeat (SRSender).
Added a new file that will contain the implementations of the
classes for selective-repeat called selectiveRepeat.js.
The implemented tests are currently not passing as SRSender
is not implemented.