Changed onReceive to accept a third parameter
In class Node, changed the onReceive callback method to accept
a third parameter isOk that should be true if the received
packet was correct and false if not.
Changed the descendant classes of Node to use the onReceive
method accordingly.
Removed onReceiveOk and onReceiveNotOk from SWReceiver.
Changed one of the tests in stopAndWaitTest.js, to use the changes
of onReceive.
Removed the definition of onReceive from the classes in
selectiveRepeat.js as now Node has the same definition.
Changed the documentation of the methods that use onReceive
to specify when isOk will be true and when it will be false.