Added the JSDoc template repo as a git submodule
Added the Braintree JSDoc template repository as a git submodule.
Updated README.md to know how to generate documentation.
Improved the documentation of the library adding @extends tags
for the descendant classes of Node and changed the @module tags
at the beginning of the modules, removing their .js extension as
it could cause confusion with the JSDoc nomenclature.