Added an initial layout for the settings card
Added a new folder /components to put the different components
of the application. Added a folder /components/SettingsCard
where the files for the settings card component are placed.
Added the css part of Fontawesome 5.15.1-web to the project, this
is under /fontawesome.
Added files index.html, main.js and main.css as main entries for
the UI of the project.
Added the initial behaviour for the settings card, without adding
any inputs yet.