Renamed the folder /js to /lib
Changed the location of mocha.html to go under /lib/tests/,
changing also the urls of the files referenced inside it.
Changed jsdoc.json to generate the documentation from the new
location.
Changed the README.md file to show the correct locations of the
referenced files.
Updated the documentation to reflect the changes in README.md.