Added layout and logic for the visualization component
Created a packet-visualization component that shows vertical packet
tracks in a row and a legend with the colors of each part of the
visualization.
Created a packet-track component for encapsulating the animation
behaviour of the packet interchange for a single sequence number.
Made minor style changes to the settings card like margins and
font-size.