Connected the simulation execution to the visualization component
Changed some initial parameters of the settings form in index.html
and their name attributes.
For the packet-track component added new methods to stop the
current timer animation and to remove packets from the visualization.
Also added a callback for handling packet clicks and improved the
startTimer and sendPacket methods.
In main.js connected the visualization to the simulation.
Improved the logic and animations of the visualization component,
and added methods for starting and stopping timeout animations.
In the protocols library added new method callbacks to node
for handling started and stopped timeouts. Also added the
calls to these callbacks to GBNSender, SWSender and SRSender.
Additionally added a new property to Packet that tells whether
the packet was re-sent or not and helps selecting the proper
track in the visualization component.
Lastly, updated the documentation.