Refactored the packet track component
Changed the default visibility of a travelin packet to hidden
and made it visible only while the animation occurs.
Improved the documenation headers of some methods.
Improved the reset() method, stopping any packet animations
that may still be running.
Changed the name of method removePacket() to losePacket(), and
improved its documentation.