Added visualization of damaged packets
Added a new form element to input the damage probability.
Added a new method damagePacket() to the packet track component
and the packet visualization component.
Added a new label to the visualization legend for the damaged packets.
Changed main.js to connect the new input parameter with the channel
options and the visualization.