Added a pause/resume button to the settings card
Added 2 new methods to the settings card pause() and resume().
Changed main.js to allow pausing/resuming of the simulation and
to automatically pause the simulation when the user switches to
another tab.
For now pausing and resuming only works for stop-and-wait.
Fixed a bug of the stop() method in SWSender.