Changed SRSender and SRReceiver to use cumulative acknowledgments
Added a new field to Packet, isCAck that is true if the packet
is a cumulative acknowledgment.
Added a boolean  option for the constructor of SRReceiver to use
cumulative acknowledgments.
Changed the mehtods receive() of both SRSender and SRReceiver.
Added a new option in the input of the settings card component
for choosing Selective Repeat + CACK.
Changed main.js to allow the new option.
Changed the packet visualization component to show previous
packets as confirmed when a packet that is a cumulative acknowledgment
arrives.